requirements.txt
customtkinter==5.2.0
docx2pdf==0.1.8
numpy==1.25.2
Pillow==10.0.0
Pillow==10.3.0
python_docx==1.1.0
tkcalendar==1.6.1
